Kafka: multi-node working! Found the bug
I was redeclaring the offset variable inside the while loop for the
retries, instead of making the variable mutable and giving it the new
value. When leaving the while loop, the new (and correct) offset
variable would go out of scope and the node would reply with the old and
incorrect value. Lesson learnt!